Change int to long

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -2,25 +2,25 @@ package ru.netology.stats;
 
 public class StatsService {
 
-    public int salesAmount(int[] monthsOfSales) {
-        int sum = 0;
-        for (int monthsOfSale : monthsOfSales) {
+    public long salesAmount(long[] monthsOfSales) {
+        long sum = 0;
+        for (long monthsOfSale : monthsOfSales) {
             sum = sum + monthsOfSale;
 
         }
         return sum;
     }
 
-    public int averageSalesAmount(int[] monthsOfSales) {
-        int sum = 0;
-        for (int monthsOfSale : monthsOfSales) {
+    public long averageSalesAmount(long[] monthsOfSales) {
+        long sum = 0;
+        for (long monthsOfSale : monthsOfSales) {
             sum = sum + monthsOfSale;
         }
         return sum / 12;
 
     }
 
-    public int maxSales(int[] monthsOfSales) {
+    public int maxSales(long[] monthsOfSales) {
         int maxMonth = 0; // номер месяца с максимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < monthsOfSales.length; i++) {
@@ -32,7 +32,7 @@ public class StatsService {
         return maxMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int minSales(int[] monthsOfSales) {
+    public int minSales(long[] monthsOfSales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < monthsOfSales.length; i++) {
@@ -44,11 +44,11 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int numberOfMonthsOfSalesBelowAverage(int[] monthsOfSales) {
-        int average = averageSalesAmount(monthsOfSales);
-        int numberOfMonths = 0; //счётчик месяцев с продажами ниже среднего
+    public long numberOfMonthsOfSalesBelowAverage(long[] monthsOfSales) {
+        long average = averageSalesAmount(monthsOfSales);
+        long numberOfMonths = 0; //счётчик месяцев с продажами ниже среднего
 
-        for (int monthsOfSale : monthsOfSales) {
+        for (long monthsOfSale : monthsOfSales) {
             if (monthsOfSale < average) {
                 numberOfMonths++;
             }
@@ -56,10 +56,10 @@ public class StatsService {
         return numberOfMonths;
     }
 
-    public int numberOfMonthsOfSalesAboveAverage(int[] monthsOfSales) {
-        int average = averageSalesAmount(monthsOfSales);
-        int numberOfMonths = 0; //счётчик месяцев с продажами выше среднего
-        for (int monthsOfSale : monthsOfSales) {
+    public long numberOfMonthsOfSalesAboveAverage(long[] monthsOfSales) {
+        long average = averageSalesAmount(monthsOfSales);
+        long numberOfMonths = 0; //счётчик месяцев с продажами выше среднего
+        for (long monthsOfSale : monthsOfSales) {
             if (monthsOfSale > average) {
                 numberOfMonths++;
             }
